+++ client/components/cart/cartMain.jsx
@@ -1,15 +1,9 @@
 import React from 'react'
 import {connect} from 'react-redux'
-import {fetchCart} from '../store/cart'
+import {fetchCart} from '../../store/cartStore'
 import {CartList} from './cartList'
-// import {userForm} from './Checkout/userForm'
 
-const defaultCart = {
-  id: 0,
-  isPending: false,
-  products: []
-}
-export class Cart extends React.Component {
+export class CartMain extends React.Component {
   async componentDidMount() {
     await this.props.getCart()
   }
@@ -19,7 +13,6 @@ export class Cart extends React.Component {
       <div>
         <h3>C A R T</h3>
         <CartList cart={this.props.cart} />
-        {/* <userForm cart={this.props.cart} /> */}
       </div>
     )
   }
@@ -33,4 +26,4 @@ const mapDispatch = dispatch => ({
   getCart: () => dispatch(fetchCart())
 })
 
-export default connect(mapState, mapDispatch)(Cart)
+export default connect(mapState, mapDispatch)(CartMain)
