+++ client/components/index.js
@@ -3,11 +3,15 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as AddProductForm} from './AddProductForm'
-export {default as AllProducts} from './AllProducts'
-export {default as Cart} from './cart'
-export {default as CartList} from './cartList'
+export {default as AddProductForm} from './addProductForm'
+export {default as AllProducts} from './allProducts'
+export {default as CartMain} from './cart/cartMain'
+export {default as CartList} from './cart/cartList'
+export {default as CartProducts} from './cart/cartProducts'
+export {default as CartQuantity} from './cart/cartQuantity'
+
 export {default as Navbar} from './navbar'
 export {default as SingleProduct} from './singleProduct'
 export {default as UserHome} from './userHome'
+export {default as UserForm} from './checkout/userForm'
 export {Login, Signup} from './authForm'
